+++ core/src/DynamicStripes.cpp
@@ -145,7 +145,7 @@ namespace core {
 
         // Stats
         auto index = stats.cycles.size();
-        stats.cycles.emplace_back(std::vector<uint32_t>(batch_size,0));
+        stats.cycles.emplace_back(std::vector<uint64_t>(batch_size,0));
 
         std::vector<int> list_x, list_y;
         int n, x_counter, y_counter;
@@ -171,7 +171,7 @@ namespace core {
             stats.cycles[index][n] *= num_filters_sets;
         }
 
-        auto base_cycles = (uint32_t)(out_x * out_y * act_channels * Kx * Ky * num_filters_sets / N_ROWS);
+        auto base_cycles = (uint64_t)(out_x * out_y * act_channels * Kx * Ky * num_filters_sets / N_ROWS);
 
         std::chrono::high_resolution_clock::time_point t2 = std::chrono::high_resolution_clock::now();
         std::chrono::duration<double> time_span = std::chrono::duration_cast<std::chrono::duration<double>>(t2 - t1);
@@ -208,7 +208,7 @@ namespace core {
 
         // Stats
         auto index = stats.cycles.size();
-        stats.cycles.emplace_back(std::vector<uint32_t>(batch_size,0));
+        stats.cycles.emplace_back(std::vector<uint64_t>(batch_size,0));
 
         int n;
 
@@ -253,7 +253,7 @@ namespace core {
 
         #endif
 
-        auto base_cycles = (uint32_t)(act_channels * num_filters_sets / N_ROWS);
+        auto base_cycles = (uint64_t)(act_channels * num_filters_sets / N_ROWS);
 
         std::chrono::high_resolution_clock::time_point t2 = std::chrono::high_resolution_clock::now();
         std::chrono::duration<double> time_span = std::chrono::duration_cast<std::chrono::duration<double>>(t2 - t1);
