+++ core/src/Laconic.cpp
@@ -122,7 +122,7 @@ namespace core {
 
         // Stats
         auto index = stats.cycles.size();
-        stats.cycles.emplace_back(std::vector<uint32_t>(batch_size,0));
+        stats.cycles.emplace_back(std::vector<uint64_t>(batch_size,0));
 
         int current_group = 0, group_m =0, start_group = 0;
         std::vector<int> list_x, list_y;
@@ -186,7 +186,7 @@ namespace core {
 
         // Stats
         auto index = stats.cycles.size();
-        stats.cycles.emplace_back(std::vector<uint32_t>(batch_size,0));
+        stats.cycles.emplace_back(std::vector<uint64_t>(batch_size,0));
 
         int n;
